requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.2
django-phone-field==1.8.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
msgpack-python==0.5.6
Pillow==7.2.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
redis==2.10.6
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.2.0
zope.interface==5.0.2
